backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26       
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12        
filelock==3.18.0
fsspec==2025.3.2
greenlet==3.2.1
h11==0.16.0
httptools==0.6.4
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==80.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
sympy==1.14.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1